requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Date Handling
python-dateutil>=2.8.0

# HTTP Requests
requests>=2.31.0

# Environment Variables
python-dotenv>=1.0.0

# Airflow (optional - for production deployment)
# apache-airflow>=2.7.0

# Jupyter Notebook
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0

# Data Visualization (for notebook)
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality (optional)
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
